Display system tags in popup
