+++ client/src/components/city/lines-tree.js
@@ -1,5 +1,6 @@
 import React, {PureComponent} from 'react';
 import Translate from 'react-translate-component';
+import SystemTags from './system-tags';
 
 class LinesTree extends PureComponent {
   constructor(props, context) {
@@ -32,7 +33,7 @@ class LinesTree extends PureComponent {
     return (
       <ul className="c-tree system-tree">
         <li className={`c-tree__item ${expandClass}`}>
-          <span className="c-link" onClick={this.toggleExpanded.bind(this)}>{this.props.system.name || <Translate content="city.lines" />}</span>
+          <span className="c-link system-name" onClick={this.toggleExpanded.bind(this)}>{this.props.system.name || <Translate content="city.lines" />}</span>
           <SystemTags system={this.props.system} />
           <ul className="c-tree" style={{display: this.state.expanded ? 'block' : 'none'}}>
             { lines.length > 1 ?
@@ -56,21 +57,6 @@ class LinesTree extends PureComponent {
           </ul>
         </li>
       </ul>
-    )
-  }
-}
-
-class SystemTags extends PureComponent {
-  tags() {
-    return ['historic', 'project'];
-  }
-
-  render() {
-    return this.tags().map(tag =>
-      this.props.system[tag] &&
-        <span key={tag} className="c-badge c-badge--ghost c-badge--brand system-tag">
-          {tag}
-        </span>
     )
   }
 }
